Move GC-related declarations to gc folder and add comments.
